% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escaped_quote.R
\name{escapedQuote}
\alias{escapedQuote}
\title{Escape quotes for shell commands}
\usage{
escapedQuote(x, double = FALSE)
}
\arguments{
\item{x}{A character vector that may contain double quotes.}

\item{double}{Boolean. Should the quotes be escaped twice? Default: FALSE}
}
\value{
A character vector with all double quotes double escaped so
that it can safely be passed to a command line call.
}
\description{
This helper prepares character vectors for use in shell commands by
quoting them with \code{shQuote()} and (double) escaping any internal
double quotes.
}
\examples{
\dontrun{
escapedQuote('foo "bar" baz')
}

}
\keyword{internal}
