% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather_package_names}
\alias{gather_package_names}
\title{Gather Package Names}
\usage{
gather_package_names()
}
\value{
A character vector of R package names.
}
\description{
Searches all meta data in the current directory for
information on what R packages are necessary to
reproduce the Markdown files.
}
