% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reproducibility_summary.R
\name{get_num_reproducibility_errors}
\alias{get_num_reproducibility_errors}
\title{Get the total number of failed reproduction attempts}
\usage{
get_num_reproducibility_errors()
}
\description{
Get the total number of failed reproduction attempts
}
