% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isReproducible.R
\name{isReproducible}
\alias{isReproducible}
\title{Test reproducibility of an R Markdown file}
\usage{
isReproducible(filename, ...)
}
\arguments{
\item{filename}{An R Markdown file to check for reproducibility}

\item{\ldots}{Optional arguments passed down to rmarkdown::render()}
}
\description{
Test reproducibility of an R Markdown file
}
