% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_repro_data.R
\name{load_repro_data}
\alias{load_repro_data}
\title{Loading reproducibility data}
\usage{
load_repro_data(filename, envir = globalenv(), filetype = c("json", "rda"))
}
\arguments{
\item{filename}{Character. Filename to load objects from.}

\item{envir}{Environment to load the objects into. By default, this is the global environment.}

\item{filetype}{Character. Currently supported is json and rda.}
}
\description{
Loading reproducibility data
}
