% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{Delete reproducibility files}
\usage{
reset(interactive = TRUE)
}
\arguments{
\item{interactive}{Logical. If TRUE, ask for confirmation before deleting.}
}
\value{
Invisibly returns the vector of deleted files.
}
\description{
Deletes all files in the current working directory that start with \code{default_prefix()} and end with \code{default_filetype()}.
}
