% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reproducibility_summary.R
\name{get_reproducibility_summary}
\alias{get_reproducibility_summary}
\title{Get a summary about all reproduction attempts}
\usage{
get_reproducibility_summary()
}
\value{
Returns a data.frame with three columns.
}
\description{
This function returns a data frame, in which details
about reproduction attempts are collected. The data frame has
three columns named "Chunk","Variable", and "Success". Every row
in the data frame corresponds to one variable, for which reproducibility
was tested. \code{Chunk} stores the name of the surrounding chunk, \code{Variable}
stores the name of the variable, and \code{Success} is a boolean variable,
which indicates whether the reproduction attempt was successful.
}
