% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibleR.R
\name{reproducibleR}
\alias{reproducibleR}
\title{This is the main RMarkdown chunk hook for processing
the reproducibility tests}
\usage{
reproducibleR(options)
}
\description{
This is the main RMarkdown chunk hook for processing
the reproducibility tests
}
