% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibleR.R
\name{reproducibleR}
\alias{reproducibleR}
\title{Knitr Hook}
\usage{
reproducibleR(options)
}
\arguments{
\item{options}{A list of chunk options passed from the knitr engine.}
}
\description{
This is the main RMarkdown chunk hook for processing
the automated reproducibility tests of code chunks
}
\details{
This function first executes the R code from a given chunk.
}
\author{
Andreas M. Brandmaier
}
