% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_repro_data.R
\name{save_repro_data}
\alias{save_repro_data}
\title{Storing reproducibility data}
\usage{
save_repro_data(x, filename, filetype = default_filetype(), envir = NULL)
}
\arguments{
\item{filename}{Name (possible including full path) of the save file}

\item{filetype}{Character. Currenlty supported is json and rda.}

\item{envir}{Environment to load the objects into. By default, this is the global environment.}
}
\description{
Storing reproducibility data
}
